% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native-raster.R
\name{as_pattern}
\alias{as_pattern}
\title{Cast a vector into a native raster}
\usage{
as_pattern(x, width, height, ...)
}
\arguments{
\item{x}{A vector to be cast into a native raster.}

\item{width, height}{Integer scalars giving the width and height of the image.}

\item{...}{Additional arguments.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Casts a vector of pixel values into a native raster.
The vector must be of length \code{width * height}
}
