% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attenuate.R
\name{attenuate}
\alias{attenuate}
\alias{bounce_off}
\title{Generate an attenuated sequence}
\usage{
attenuate(n, init, velocity, damping = 1)

bounce_off(n, init, velocity, limits, damping = -1, mass = 0)
}
\arguments{
\item{n}{An integer scalar. The number of steps in the sequence.}

\item{init}{A numeric scalar. The initial value of the sequence.}

\item{velocity}{A numeric scalar. The velocity or change rate for each step.}

\item{damping}{A numeric scalar.
The damping factor applied to the velocity at each step.}

\item{limits}{A numeric vector of length 2.
The minimum and maximum values allowed in the sequence.}

\item{mass}{A numeric scalar. The mass of the object bouncing.}
}
\value{
A numeric vector.
}
\description{
Generates a sequence of values starting from an initial value,
with a given velocity that is attenuated over time.

For \code{bounce_off()}, a sequence is constrained
so that the values never go outside of a given range.
The sequence is generated by applying a damping factor
to the velocity when it exceeds \code{limits}.
}
