% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-spiral.R
\name{curve_archimedean}
\alias{curve_archimedean}
\title{Generate an Archimedean spiral}
\usage{
curve_archimedean(n, a = 0, b = 1, c = 1, base = exp(1))
}
\arguments{
\item{n}{An integer scalar; Number of points to sample along the curve.}

\item{a, b, c}{Numeric scalars; Parameters of the general Archimedean spiral.}

\item{base}{A numeric scalar;
Base of the logarithm used to compute the spacing between points.}
}
\value{
A data frame with columns \code{x} and \code{y}.
}
\description{
Generates a 2D curve that is a general Archimedean spiral
where the raidus \code{r} is defined as follows:
\deqn{
r = a + b \cdot \theta^\frac{1}{c}
}
}
\details{
According to the \href{https://en.wikipedia.org/wiki/Archimedean_spiral}{Wikipedia article},
the normal Archimedean spiral occurs when \code{c = 1}.
Other spirals falling into this group include
the hyperbolic spiral (\code{c = -1}),
Fermat's spiral (\code{c = 2}), and the lituus (\verb{c = −2}).
}
\seealso{
Other curve: 
\code{\link{bezier}},
\code{\link{curve_involute}()},
\code{\link{curve_spring}()}
}
\concept{curve}
