% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-random.R
\name{curve_random}
\alias{curve_random}
\title{Generate a random curve}
\usage{
curve_random(n, power = 8, hurst_index = 0.5, smoothing = 0.4, ...)
}
\arguments{
\item{n}{An integer scalar. The number of points to generate.}

\item{power}{Integer.}

\item{hurst_index}{Numeric in range \verb{(0, 1)}.}

\item{smoothing}{An numeric scalar.
Smoothing parameter to be passed to \code{\link[stats:loess]{stats::loess()}} as \code{span}.}

\item{...}{Additional arguments to be passed to \code{\link[stats:loess]{stats::loess()}}.}
}
\value{
A data frame with columns \code{x} and \code{y}.
}
\description{
Generates a random smooth curve using fractional Brownian bridges.
This function fits a series of points that are generated by \code{\link[=fbbridge_from]{fbbridge_from()}}
using \code{\link[stats:loess]{stats::loess()}} and returns predicted points for \code{n} time steps.
}
\seealso{
Other curve: 
\code{\link{curve_archimedean}()},
\code{\link{curve_involute}()},
\code{\link{curve_spring}()},
\code{\link{curve_twist}()}
}
\concept{curve}
