% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{deg-rad}
\alias{deg-rad}
\alias{deg2rad}
\alias{rad2deg}
\title{Convert degrees and radians}
\usage{
deg2rad(deg)

rad2deg(rad)
}
\arguments{
\item{deg}{Degrees.}

\item{rad}{Radians.}
}
\value{
A numeric vector.
}
\description{
Convert degrees and radians
}
