% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{delaunay}
\alias{delaunay}
\title{Compute Delaunay triangulation}
\usage{
delaunay(seeds, x = x, y = y)
}
\arguments{
\item{seeds}{A data frame containing the input points.}

\item{x, y}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns in \code{seeds}
that contain the x- and y-coordinates.}
}
\value{
A list with two data frames:
\itemize{
\item \code{circumcenters}: A tibble containing the triangle id, the x- and
y-coordinates of the circumcenter, and the circumradius.
\item \code{vertices}: A tibble containing the triangle id and the x- and
y-coordinates of the three vertices of each triangle.
}
}
\description{
Computes the Delaunay triangulation of a set of 2D points.
}
\details{
This function returns the triangulation in a form that is convenient for
drawing. \code{vertices} stores the triangle vertices in long format, with three
rows per triangle. \code{circumcenters} stores the corresponding circumcircle of
each triangle.
}
\examples{
theta <- seq(-pi, pi, length.out = 12)
seeds <- data.frame(
  x = c(150 * cos(theta), 0),
  y = c(150 * sin(theta), 0)
)

tri <- delaunay(seeds)

tri$circumcenters
tri$vertices

}
