% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{expand}
\alias{expand}
\title{Expand grid}
\usage{
expand(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[=expand.grid]{expand.grid()}}.}
}
\value{
A tibble.
}
\description{
A thin wrapper for \code{\link[=expand.grid]{expand.grid()}} that returns a tibble
while converting numeric columns to double.
}
