% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm.R
\name{fbbridge_from}
\alias{fbbridge_from}
\title{Create a 1D fractional Brownian bridge generator}
\usage{
fbbridge_from(power = 4, hurst_index = 0.5)
}
\arguments{
\item{power}{Integer.}

\item{hurst_index}{Numeric in range \verb{(0, 1)}.}
}
\value{
A function that takes a function \code{func} as its first argument
and returns a time series.
\code{func} is expected to take \code{n} as its first argument
and return a numeric vector of length \code{n}.
}
\description{
Create a 1D fractional Brownian bridge generator
}
\examples{
fbbridge_from()(runif, min = -1, max = 1)
}
