% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm.R
\name{fbm}
\alias{fbm}
\alias{fbm_from}
\alias{fbbridge_from}
\alias{fbbridge_2d_from}
\title{Generators for FBM and FBB}
\usage{
fbm_from(power = 4, hurst_index = 0.5)

fbbridge_from(power = 4, hurst_index = 0.5)

fbbridge_2d_from(power = 4, hurst_index = 0.5)
}
\arguments{
\item{power}{Integer.}

\item{hurst_index}{Numeric in range \verb{(0, 1)}.}
}
\value{
A function that takes a function \code{func} as its first argument
and returns a time series (for \code{fbm_from()} and \code{fbbridge_from()})
or a matrix (for \code{fbm_2d_from()}).
\code{func} is expected to take \code{n} as its first argument
and return a numeric vector of length \code{n}.
}
\description{
Creates a function that generates a fractional Brownian motion (FBM) or
fractional Brownian bridge (FBB) time series with a given Hurst index.
}
