% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{\%!*\%}
\alias{\%!*\%}
\title{NDC multiplication}
\usage{
lhs \%!*\% rhs
}
\arguments{
\item{lhs}{A \code{transform3d} object.}

\item{rhs}{A \code{transform3d} object.}
}
\value{
A \code{transform3d} object.
}
\description{
NDC multiplication
}
\keyword{internal}
