% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pingpong}
\alias{pingpong}
\title{Ping-pong sequence}
\usage{
pingpong(x, ...)
}
\arguments{
\item{x}{An object to repeat.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Repeats \code{x} in a ping-pong fashion.
}
