% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{seq_ease}
\alias{seq_ease}
\alias{seq_color}
\title{Interpolate between two values}
\usage{
seq_ease(x1, x2, n = 100, ease = function(t) ease_in_out(t, "cubic"))

seq_color(
  col1,
  col2,
  n = 100,
  ease = function(t) ease_in_out(t, "cubic"),
  colorspace = "hcl"
)
}
\arguments{
\item{x1, x2}{Numeric scalars.}

\item{n}{\code{length.out} parameter for \code{\link[base:seq]{base::seq()}}.}

\item{ease}{A function to ease the sequence.}

\item{col1, col2}{Character scalars. The colors to interpolate between.}

\item{colorspace}{A string. Color space in which to do the interpolation.
See \code{\link[farver:convert_colour]{farver::convert_colour()}} for details.}
}
\value{
\itemize{
\item For \code{seq_ease()}, a numeric vector.
\item For \code{seq_color()}, a character vector.
}
}
\description{
Derived from
\href{https://github.com/coolbutuseless/displease}{coolbutuseless/displease}.
}
