% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smoothstep}
\alias{smoothstep}
\alias{smootherstep}
\alias{ease_in}
\alias{ease_out}
\alias{ease_in_out}
\title{Smoothing functions}
\usage{
smoothstep(t)

smootherstep(t)

ease_in(
  t,
  type = c("sine", "quad", "cubic", "quart", "quint", "exp", "circle", "elastic", "back",
    "bounce")
)

ease_out(
  t,
  type = c("sine", "quad", "cubic", "quart", "quint", "exp", "circle", "elastic", "back",
    "bounce")
)

ease_in_out(
  t,
  type = c("sine", "quad", "cubic", "quart", "quint", "exp", "circle", "elastic", "back",
    "bounce")
)
}
\arguments{
\item{t}{A numeric vector.}

\item{type}{A string. The type of easing function to use.}
}
\value{
A numeric vector.
}
\description{
Smoothing functions
}
