% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-bouncing-pts.R
\name{state_bouncing_pts}
\alias{state_bouncing_pts}
\title{Generate a state of bouncing points}
\usage{
state_bouncing_pts(
  n,
  bbox = c(-1, -1, 1, 1),
  speed = 1,
  restitution = 1,
  seed = NULL
)
}
\arguments{
\item{n}{An integer scalar; Number of bouncing points.}

\item{bbox}{Numeric vector of length 4;
The bounding box of the bouncing points as \code{c(xmin, ymin, xmax, ymax)}.}

\item{speed}{Numeric scalar; The speed factor of the bouncing points.}

\item{restitution}{Numeric scalar; The restitution of the bouncing points.}

\item{seed}{Numeric scalar; Random seed for reproducibility.}
}
\value{
An external pointer that represents the state of \code{n} bouncing points.
}
\description{
Generates a state of \code{n} bouncing points.
}
