% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{camera}
\alias{camera}
\alias{lookat3d}
\alias{persp3d}
\alias{viewport3d}
\title{3D world to camera transformation}
\usage{
lookat3d(eye, center, up = c(0, 1, 0))

persp3d(fovy, aspect, near = 0.1, far = 10)

viewport3d(width, height)
}
\arguments{
\item{eye}{A numeric vector of length 3 giving the position of the camera.}

\item{center}{A numeric vector of length 3 giving the position of the target.}

\item{up}{A numeric vector of length 3 giving the direction of the "up" vector.}

\item{fovy}{A numeric scalar giving the field of view in radians.}

\item{aspect}{A numeric scalar giving the aspect ratio.}

\item{near}{A numeric scalar giving the distance to the near plane.}

\item{far}{A numeric scalar giving the distance to the far plane.}

\item{width, height}{A numeric scalar giving the width and height of the viewport.}
}
\value{
A \code{transform3d} object.
}
\description{
3D world to camera transformation
}
