% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-others.R
\name{curve-others}
\alias{curve-others}
\alias{curve_archimedean}
\alias{curve_cyclic_harmonic}
\alias{curve_epicycloid}
\alias{curve_epitrochoid}
\alias{curve_gear}
\alias{curve_heart}
\alias{curve_hypocycloid}
\alias{curve_hypotrochoid}
\alias{curve_lissajous}
\alias{curve_ranunculoid}
\alias{curve_rose}
\alias{curve_spirograph}
\title{Generate geometric curves}
\usage{
curve_archimedean(n, a = 0, b = 1, c = 1, base = exp(1))

curve_cyclic_harmonic(n, k = 5, a = 1, b = 0.5, scale = 1, base = exp(1))

curve_epicycloid(n, or = 16, ir = 3.5, scale = 1, base = exp(1))

curve_epitrochoid(n, or = 16, ir = 3.5, b = 2.4, scale = 1, base = exp(1))

curve_gear(n, k = 10, a = 1, b = 16, scale = 1, base = exp(1))

curve_heart(n, scale = -1)

curve_hypocycloid(n, or = 16, ir = 3.5, scale = 1, base = exp(1))

curve_hypotrochoid(n, or = 16, ir = 3.5, b = 2.4, scale = 1, base = exp(1))

curve_lissajous(n, d = 10, e = 16, scale = 1, base = exp(1))

curve_ranunculoid(n, k = 6, scale = 1/k, base = exp(1))

curve_rose(n, k = 5, c = 1, scale = 1, base = exp(1))

curve_spirograph(n, or = 16, ir = 3.5, b = 2.4, scale = 1, base = exp(1))
}
\arguments{
\item{n}{An integer scalar; Number of points to sample along the curve.}

\item{a, b, c, d, e, k}{Numeric scalars; Parameters of the curve.}

\item{base}{A numeric scalar;
Base of the logarithm used to compute the spacing between points.}

\item{scale}{A numeric scalar; Scaling factor for the curve.}

\item{or, ir}{Numeric scalars; Outer and inner radius.}
}
\value{
\itemize{
\item For polar equations, a tibble with columns \code{id}, \code{phi}, \code{r}, \code{x}, and \code{y}.
\item For others, a tibble with columns \code{id}, \code{theta}, \code{x}, and \code{y}.
}
}
\description{
Generates a data frame of points along a geometric curve.

\code{"curve"} in rasengan is a generic term for a set of functions
that take parameters \code{n} and generate a data frame
with \code{n} rows and columns containing \code{x} and \code{y} along the geometric curve.
\subsection{Archimedean spirals}{

\code{curve_archimedean} generates a 2D curve that is a general Archimedean spiral
where the raidus \code{r} is defined as follows:
\deqn{
r = a + b \cdot \theta^\frac{1}{c}
}

According to the \href{https://en.wikipedia.org/wiki/Archimedean_spiral}{Wikipedia article},
the normal Archimedean spiral occurs when \code{c = 1}.
Other spirals falling into this group include
the hyperbolic spiral (\code{c = -1}),
Fermat's spiral (\code{c = 2}), and the lituus (\verb{c = −2}).
}
}
\seealso{
Other curve:
\code{\link{bezier}}
}
\concept{curve}
