% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-spiral.R
\name{curve_involute}
\alias{curve_involute}
\title{Generate an involute}
\usage{
curve_involute(n, scale = 1, base = exp(1))
}
\arguments{
\item{n}{An integer scalar; Number of points to sample along the curve.}

\item{scale}{A numeric scalar; Scaling factor for the curve.}

\item{base}{A numeric scalar;
Base of the logarithm used to compute the spacing between points.}
}
\value{
A data frame with columns \code{x} and \code{y}.
}
\description{
Generates a 2D curve that is an involute curve.
}
\seealso{
Other curve:
\code{\link{bezier}},
\code{\link[=curve_archimedean]{curve_archimedean()}},
\code{\link[=curve_spring]{curve_spring()}}
}
\concept{curve}
