% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-spiral.R
\name{curve_spring}
\alias{curve_spring}
\title{Generate a balance spring curve}
\usage{
curve_spring(n, k = 1, m = 0.1, scale = 1, base = exp(1))
}
\arguments{
\item{n}{An integer scalar; Number of points to sample along the curve.}

\item{k, m}{Numeric scalars; Parameters of the balance spring curve.}

\item{scale}{A numeric scalar; Scaling factor for the curve.}

\item{base}{A numeric scalar; Base of the logarithm used to compute the spacing between points.}
}
\value{
A data frame with columns \code{x} and \code{y}.
}
\description{
Generates a 2D curve that is a balance spring curve.
}
\seealso{
Other curve: 
\code{\link{curve_archimedean}()},
\code{\link{curve_involute}()},
\code{\link{curve_random}()},
\code{\link{curve_twist}()}
}
\concept{curve}
