% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{domain_warp}
\alias{domain_warp}
\title{Generate data with domain warping}
\usage{
domain_warp(
  data,
  seed = sample.int(1337, 1),
  warp_type = c("OpenSimplex2", "OpenSimplex2Reduced", "BasicGrid"),
  amplitude = 1,
  fractal_type = c("None", "DomainWarpProgressive", "DomainWarpIndependent"),
  octaves = 3,
  lacunarity = 2,
  gain = 0.5,
  rotation_type = c("None", "ImproveXYPlanes", "ImproveXZPlanes")
)
}
\arguments{
\item{data}{A numeric matrix that has just 2 or 3 columns.}

\item{seed}{An integer scalar; Random seed.}

\item{warp_type}{A string; Warp type.}

\item{amplitude}{A numeric scalar; Amplitude (the maximum warp distance from original position).}

\item{fractal_type}{A string; Fractal type.}

\item{octaves}{A numeric scalar; Number of octaves.}

\item{lacunarity}{A numeric scalar; Lacunarity (the frequency multiplier between each octave).}

\item{gain}{A numeric scalar; Gain (the relative strength of noise from each layer when compared to the last).}

\item{rotation_type}{A string; Rotation type for 3D noise.}
}
\value{
A double matrix.
}
\description{
Warps data with domain warping.
}
\examples{
if (requireNamespace("dplyr", quietly = TRUE)) {
 nz <- expand(
   id = seq(0, 1, length.out = 16),
   x = seq(0, 32, by = 2),
   y = seq(0, 32, by = 2)
 ) |>
   dplyr::mutate(
     val = noise_3d()(data = cbind(id, x, y) |> domain_warp()),
     id = dplyr::consecutive_id(id)
   )
}
if (require("ggplot2", quietly = TRUE)) {
  ggplot(nz) +
    geom_tile(aes(x = x, y = y, fill = val))
}
}
