% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{misc}
\alias{misc}
\alias{deg2rad}
\alias{rad2deg}
\alias{fract}
\alias{mag}
\alias{pingpong}
\title{Miscellaneous functions}
\usage{
deg2rad(x)

rad2deg(x)

fract(x)

mag(mat, origin = c(0, 0))

pingpong(x)
}
\arguments{
\item{x}{A numeric vector.}

\item{mat}{A numeric matrix or a data frame.}

\item{origin}{A numeric vector to be subtracted from \code{mat}.}
}
\value{
A numeric vector.
}
\description{
Miscellaneous functions
}
