% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{misc}
\alias{misc}
\alias{deg2rad}
\alias{rad2deg}
\alias{fract}
\alias{mag}
\title{Miscellaneous functions}
\usage{
deg2rad(x)

rad2deg(x)

fract(x)

mag(x, origin = c(0, 0))
}
\arguments{
\item{x}{A numeric vector.}

\item{origin}{A numeric vector of length 2.}
}
\value{
A numeric vector.
}
\description{
Miscellaneous functions
}
