% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modifications}
\alias{modifications}
\alias{blend}
\alias{normalise}
\alias{normalize}
\alias{cap}
\alias{pulse}
\title{Simply value modifications}
\usage{
blend(x, y, mask)

normalise(x)

normalize(x, from = range(x), to = c(0, 1))

cap(x, lower = 0, upper = 1)

pulse(x, mask)
}
\arguments{
\item{x, y}{A numeric vector.}

\item{mask}{A numeric scalar, typically between \code{0} and \code{1}.}

\item{from}{A numeric vector of length 2.
The range of \code{x} to use for normalization.}

\item{to}{A numeric vector of length 2. The output domain to normalize to.}

\item{lower, upper}{A numeric scalar. The lower and upper bounds to cap to.}
}
\value{
A numeric vector.
}
\description{
Some of these functions are copied
from the \href{https://github.com/thomasp85/ambient}{ambient} package.
}
