% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{ndc_mul}
\alias{ndc_mul}
\alias{\%!*\%}
\title{Perspective division}
\usage{
ndc_mul(lhs, rhs)

lhs \%!*\% rhs
}
\arguments{
\item{lhs, rhs}{A numeric matrix.}
}
\value{
A numeric matrix.
}
\description{
Multiplication of two matrices after normalizing the first one.
}
\seealso{
camera
}
