% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-clothiod.R
\name{path_clothoid}
\alias{path_clothoid}
\title{Generate an Euler spiral or biarc curve path}
\usage{
path_clothoid(
  start = c(0, 0, pi/4),
  end = c(25, 10, pi),
  max_n = 100,
  max_iter_num = 1000,
  biarch = TRUE
)
}
\arguments{
\item{start}{A numeric vector of length 3
specifying the starting point and heading direction:
\code{c(x, y, theta)}, where \code{theta} is the orientation angle in radians.}

\item{end}{A numeric vector of length 3
specifying the goal point and heading direction:
\code{c(x, y, theta)}, where \code{theta} is the orientation angle in radians.}

\item{max_n}{A numric scalar;
Maximum number of points to sample along the path.}

\item{max_iter_num}{A numric scalar;
Maximum number of iterations used in the internal parameter estimation algorithm
(used only when \code{biarch = FALSE}).}

\item{biarch}{A logical scalar; If \code{TRUE},
use a biarc approximation consisting of two Euler spiral segments.
If \code{FALSE}, compute a single spiral curve via numerical estimation.}
}
\value{
A data frame with columns \code{x}, \code{y}, and \code{theta},
giving the sampled positions and orientations along the generated path.
If the parameter estimation fails,
the returned data may contain missing values, which are automatically removed.
}
\description{
Computes a path connecting two points, each with a specified orientation angle.
Two methods are available: a single Euler spiral, or a biarc composed of two spiral segments.
}
\details{
When \code{biarch = TRUE}, the path is constructed from two spiral segments computed using
an internal algorithm that attempts to connect the start and goal points with \( G^1 \) continuity.
This approach may be more robust in cases where a single spiral does not converge well.

When \code{biarch = FALSE}, the function attempts to compute a single Euler spiral that smoothly
interpolates between the two endpoints and their respective directions. This may fail to converge
in some configurations.
}
\examples{
\dontrun{
path <- path_clothoid()
with(path, plot(x, y, type = "l", asp = 1))
}
}
\seealso{
\href{https://github.com/CoffeeKumazaki/euler_spiral}{CoffeeKumazaki/euler_spiral: The c++ implementation for Euler spiral (clothoid).}

Other path: 
\code{\link{path_mouse}()}
}
\concept{path}
