% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasengan-package.R
\docType{package}
\name{rasengan-package}
\alias{rasengan}
\alias{rasengan-package}
\title{rasengan: Generation of Geometric Curves}
\description{
Provides functions to generate and sample geometric curves. Each function returns a data frame of 2D coordinates, suitable for visualization or further geometric processing.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Akiru Kato \email{paithiov909@gmail.com}
}

Other contributors:
\itemize{
  \item Jordan Peck [copyright holder]
  \item Kjetil Olsen Lye [copyright holder]
  \item Kohei Kumazaki [copyright holder]
  \item Øystein Myrmo [copyright holder]
  \item Wollnashorn [copyright holder]
  \item Google Inc. [copyright holder]
}

}
\keyword{internal}
