% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasengan-package.R
\docType{package}
\name{rasengan-package}
\alias{rasengan}
\alias{rasengan-package}
\title{rasengan: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
