% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-bouncing-pts.R
\name{state_bouncing_pts}
\alias{state_bouncing_pts}
\title{Generate a state for bouncing points}
\usage{
state_bouncing_pts(
  seeds,
  bbox = c(-1, -1, 1, 1),
  restitution = 1,
  x = x,
  y = y,
  vx = vx,
  vy = vy
)
}
\arguments{
\item{seeds}{A data frame containing the initial point positions.}

\item{bbox}{Numeric vector of length 4;
The bounding box of the bouncing points as \code{c(xmin, ymin, xmax, ymax)}.}

\item{restitution}{Numeric scalar; The restitution of the bouncing points.}

\item{x, y, vx, vy}{<\code{\link[rlang:args_data_masking]{data-masking}}>
Expressions specifying data columns for the initial state.}
}
\value{
An external pointer.
}
\description{
Generates a chakra state for bouncing points.
}
\examples{
seeds <-
 dplyr::tibble(
   x = runif(10, -1, 1),
   y = runif(10, -1, 1),
   angle = runif(10, -pi, pi),
   vx = cos(angle),
   vy = sin(angle)
 )

state <- state_bouncing_pts(seeds)

observe(state)

circulate(state) |>
 observe()
}
\seealso{
chakra

Other state:
\code{\link[=state_orbits]{state_orbits()}}
}
\concept{state}
